extract parseIcon from parseCommon
