+++ lib/default-profile.js
@@ -1,6 +1,9 @@
 'use strict'
 
 const parseDateTime = require('../parse/date-time')
+const parsePlatform = require('../parse/platform')
+const parseIcon = require('../parse/icon')
+const parseWhen = require('../parse/when')
 const parseDeparture = require('../parse/departure')
 const parseArrival = require('../parse/arrival')
 const parseJourneyLeg = require('../parse/journey-leg')
@@ -40,6 +43,9 @@ const defaultProfile = {
 	transformJourneysQuery: id,
 
 	parseDateTime,
+	parsePlatform,
+	parseIcon,
+	parseWhen,
 	parseDeparture,
 	parseArrival,
 	parseJourneyLeg,
